@@ -1,16 +1,17 @@
 {
   "name": "tileserver-gl",
-  "version": "3.1.1",
+  "version": "4.1.1",
   "description": "Map tile server for JSON GL styles - vector and server side generated raster tiles",
   "main": "src/main.js",
   "bin": "src/main.js",
+  "type": "module",
   "repository": {
     "type": "git",
-    "url": "https://github.com/maptiler/tileserver-gl.git"
+    "url": "git+https://github.com/maptiler/tileserver-gl.git"
   },
   "license": "BSD-2-Clause",
   "engines": {
-    "node": ">=10 <11"
+    "node": ">=14.15.0 <17"
   },
   "scripts": {
     "test": "mocha test/**.js --timeout 10000",
@@ -18,32 +19,31 @@
   },
   "dependencies": {
     "@mapbox/glyph-pbf-composite": "0.0.3",
-    "@mapbox/mapbox-gl-native": "5.0.2",
-    "@mapbox/mapbox-gl-style-spec": "13.12.0",
-    "@mapbox/mbtiles": "0.11.0",
-    "@mapbox/sphericalmercator": "1.1.0",
+    "@mapbox/mbtiles": "0.12.1",
+    "@mapbox/sphericalmercator": "1.2.0",
     "@mapbox/vector-tile": "1.3.1",
+    "@maplibre/maplibre-gl-native": "5.0.1-pre.6",
+    "@maplibre/maplibre-gl-style-spec": "17.0.1",
     "advanced-pool": "0.3.3",
-    "canvas": "2.6.1",
-    "chokidar": "3.3.1",
+    "canvas": "2.10.1",
+    "chokidar": "3.5.3",
     "clone": "2.1.2",
-    "color": "3.1.2",
-    "commander": "4.1.1",
+    "color": "4.2.3",
+    "commander": "9.4.0",
     "cors": "2.8.5",
-    "esm": "3.2.25",
-    "express": "4.17.1",
-    "handlebars": "4.7.3",
+    "express": "4.18.1",
+    "handlebars": "4.7.7",
     "http-shutdown": "1.2.2",
-    "morgan": "1.9.1",
+    "morgan": "1.10.0",
     "pbf": "3.2.1",
-    "proj4": "2.6.0",
+    "proj4": "2.8.0",
     "request": "2.88.2",
-    "sharp": "0.26.2",
+    "sharp": "0.31.0",
     "tileserver-gl-styles": "2.0.0"
   },
   "devDependencies": {
-    "mocha": "^7.1.0",
-    "should": "^13.2.3",
-    "supertest": "^4.0.2"
+    "chai": "4.3.6",
+    "mocha": "^10.0.0",
+    "supertest": "^6.2.4"
   }
 }
