@@ -1,14 +1,16 @@
 process.env.NODE_ENV = 'test';
 
-global.should = require('should');
-global.supertest = require('supertest');
+import {expect} from 'chai';
+import supertest from 'supertest';
+import {server} from '../src/server.js';
 
-require = require('esm')(module);
+global.expect = expect;
+global.supertest = supertest;
 
 before(function() {
   console.log('global setup');
   process.chdir('test_data');
-  var running = require('../src/server')({
+  const running = server({
     configPath: 'config.json',
     port: 8888,
     publicUrl: '/test/'
@@ -20,5 +22,7 @@ before(function() {
 
 after(function() {
   console.log('global teardown');
-  global.server.close(function() { console.log('Done'); process.exit(); });
+  global.server.close(function() {
+    console.log('Done'); process.exit();
+  });
 });
