@@ -1,38 +1,38 @@
-var testTileJSONArray = function(url) {
+const testTileJSONArray = function(url) {
   describe(url + ' is array of TileJSONs', function() {
     it('is json', function(done) {
       supertest(app)
-        .get(url)
-        .expect(200)
-        .expect('Content-Type', /application\/json/, done);
+          .get(url)
+          .expect(200)
+          .expect('Content-Type', /application\/json/, done);
     });
 
     it('is non-empty array', function(done) {
       supertest(app)
-        .get(url)
-        .expect(function(res) {
-          res.body.should.be.Array();
-          res.body.length.should.be.greaterThan(0);
-        }).end(done);
+          .get(url)
+          .expect(function(res) {
+            expect(res.body).to.be.a('array');
+            expect(res.body.length).to.be.greaterThan(0);
+          }).end(done);
     });
   });
 };
 
-var testTileJSON = function(url) {
+const testTileJSON = function(url) {
   describe(url + ' is TileJSON', function() {
     it('is json', function(done) {
       supertest(app)
-        .get(url)
-        .expect(200)
-        .expect('Content-Type', /application\/json/, done);
+          .get(url)
+          .expect(200)
+          .expect('Content-Type', /application\/json/, done);
     });
 
     it('has valid tiles', function(done) {
       supertest(app)
-        .get(url)
-        .expect(function(res) {
-          res.body.tiles.length.should.be.greaterThan(0);
-        }).end(done);
+          .get(url)
+          .expect(function(res) {
+            expect(res.body.tiles.length).to.be.greaterThan(0);
+          }).end(done);
     });
   });
 };
@@ -41,8 +41,8 @@ describe('Metadata', function() {
   describe('/health', function() {
     it('returns 200', function(done) {
       supertest(app)
-        .get('/health')
-        .expect(200, done);
+          .get('/health')
+          .expect(200, done);
     });
   });
 
@@ -53,21 +53,21 @@ describe('Metadata', function() {
   describe('/styles.json is valid array', function() {
     it('is json', function(done) {
       supertest(app)
-        .get('/styles.json')
-        .expect(200)
-        .expect('Content-Type', /application\/json/, done);
+          .get('/styles.json')
+          .expect(200)
+          .expect('Content-Type', /application\/json/, done);
     });
 
     it('contains valid item', function(done) {
       supertest(app)
-        .get('/styles.json')
-        .expect(function(res) {
-          res.body.should.be.Array();
-          res.body.length.should.be.greaterThan(0);
-          res.body[0].version.should.equal(8);
-          res.body[0].id.should.be.String();
-          res.body[0].name.should.be.String();
-        }).end(done);
+          .get('/styles.json')
+          .expect(function(res) {
+            expect(res.body).to.be.a('array');
+            expect(res.body.length).to.be.greaterThan(0);
+            expect(res.body[0].version).to.be.equal(8);
+            expect(res.body[0].id).to.be.a('string');
+            expect(res.body[0].name).to.be.a('string');
+          }).end(done);
     });
   });
 
